+++ tracking/TrackerUtils.py
@@ -1,0 +1,116 @@
+# Local file imports
+# TODO check what can be removed
+import numpy as np
+import os
+import sys
+import time
+import argparse
+import json
+from PIL import Image
+import matplotlib.pyplot as plt
+
+# Torch imports
+import torch
+import torch.utils.data as data
+import torch.optim as optim
+from torch.autograd import Variable
+
+
+sys.path.insert(0,'../modules')
+from sample_generator import *
+from data_prov import *
+from model import *
+from bbreg import *
+from options import *
+from gen_config import *
+
+def forward_samples(model, image, samples, out_layer='conv3'):
+    model.eval()
+    extractor = RegionExtractor(image, samples, opts['img_size'], opts['padding'], opts['batch_test'])
+    for i, regions in enumerate(extractor):
+        regions = Variable(regions)
+        if opts['use_gpu']:
+            regions = regions.cuda()
+        feat = model(regions, out_layer=out_layer)
+        if i==0:
+            feats = feat.data.clone() #somehow this f*cks up if there are certian groundtruth_rect.txt coordinates given (mine didn't work, TODO investigate this later)
+        else:
+            feats = torch.cat((feats,feat.data.clone()),0)
+    return feats
+
+
+def set_optimizer(model, lr_base, lr_mult=opts['lr_mult'], momentum=opts['momentum'], w_decay=opts['w_decay']):
+    params = model.get_learnable_params()
+    param_list = []
+    for k, p in params.iteritems():
+        lr = lr_base
+        for l, m in lr_mult.iteritems():
+            if k.startswith(l):
+                lr = lr_base * m
+        param_list.append({'params': [p], 'lr':lr})
+    optimizer = optim.SGD(param_list, lr = lr, momentum=momentum, weight_decay=w_decay)
+    return optimizer
+
+# Trains the network
+def train(model, criterion, optimizer, pos_feats, neg_feats, maxiter, in_layer='fc4'):
+    model.train()
+
+    batch_pos = opts['batch_pos']
+    batch_neg = opts['batch_neg']
+    batch_test = opts['batch_test']
+    batch_neg_cand = max(opts['batch_neg_cand'], batch_neg)
+    # TODO check how many iterations are computationally feasible for a lower loss
+    pos_idx = np.random.permutation(pos_feats.size(0))
+    neg_idx = np.random.permutation(neg_feats.size(0))
+    while(len(pos_idx) < batch_pos*maxiter):
+        pos_idx = np.concatenate([pos_idx, np.random.permutation(pos_feats.size(0))])
+    while(len(neg_idx) < batch_neg_cand*maxiter):
+        neg_idx = np.concatenate([neg_idx, np.random.permutation(neg_feats.size(0))])
+    pos_pointer = 0
+    neg_pointer = 0
+
+    for iter in range(maxiter):
+
+        # select pos idx
+        pos_next = pos_pointer+batch_pos
+        pos_cur_idx = pos_idx[pos_pointer:pos_next]
+        pos_cur_idx = pos_feats.new(pos_cur_idx).long()
+        pos_pointer = pos_next
+
+        # select neg idx
+        neg_next = neg_pointer+batch_neg_cand
+        neg_cur_idx = neg_idx[neg_pointer:neg_next]
+        neg_cur_idx = neg_feats.new(neg_cur_idx).long()
+        neg_pointer = neg_next
+
+        # create batch
+        batch_pos_feats = Variable(pos_feats.index_select(0, pos_cur_idx))
+        batch_neg_feats = Variable(neg_feats.index_select(0, neg_cur_idx))
+
+        # hard negative mining
+        if batch_neg_cand > batch_neg:
+            model.eval()
+            for start in range(0,batch_neg_cand,batch_test):
+                end = min(start+batch_test,batch_neg_cand)
+                score = model(batch_neg_feats[start:end], in_layer=in_layer)
+                if start==0:
+                    neg_cand_score = score.data[:,1].clone()
+                else:
+                    neg_cand_score = torch.cat((neg_cand_score, score.data[:,1].clone()),0)
+
+            _, top_idx = neg_cand_score.topk(batch_neg)
+            batch_neg_feats = batch_neg_feats.index_select(0, Variable(top_idx))
+            model.train()
+
+        # forward
+        pos_score = model(batch_pos_feats, in_layer=in_layer)
+        neg_score = model(batch_neg_feats, in_layer=in_layer)
+
+        # optimize
+        loss = criterion(pos_score, neg_score)
+        model.zero_grad()
+        loss.backward()
+        torch.nn.utils.clip_grad_norm(model.parameters(), opts['grad_clip'])
+        optimizer.step()
+
+        print("Iteration: %d, Loss: %.4f" % (iter, loss.data[0])) #Printing training iteration and loss